+++ src/ua/data_types/expanded_node_id.rs
@@ -10,6 +10,38 @@ use crate::{DataType as _, Error, ua};
 crate::data_type!(ExpandedNodeId);
 
 impl ExpandedNodeId {
+    /// Creates expanded node ID.
+    ///
+    /// Constructs a new instance from parts by taking ownership.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use open62541::ua;
+    ///
+    /// let node_id = ua::NodeId::numeric(1, 234);
+    /// let namespace_uri = ua::String::new("http://www.example.com").unwrap();
+    /// let server_index = 99;
+    ///
+    /// let expanded_node_id = ua::ExpandedNodeId::new(
+    ///     node_id.clone(),
+    ///     namespace_uri.clone(),
+    ///     server_index);
+    ///
+    /// assert_eq!(expanded_node_id.node_id(), &node_id);
+    /// assert_eq!(expanded_node_id.namespace_uri(), &namespace_uri);
+    /// assert_eq!(expanded_node_id.server_index(), server_index);
+    /// ```
+    #[must_use]
+    pub fn new(node_id: ua::NodeId, namespace_uri: ua::String, server_index: u32) -> Self {
+        let mut new = Self::from(node_id);
+        debug_assert!(new.namespace_uri().is_null());
+        // This passes ownership into the created expanded node ID.
+        new.0.namespaceUri = namespace_uri.into_raw();
+        new.0.serverIndex = server_index;
+        new
+    }
+
     /// Creates numeric expanded node ID.
     #[must_use]
     pub fn numeric(ns_index: u16, numeric: u32) -> Self {
@@ -21,13 +53,6 @@ impl ExpandedNodeId {
         );
 
         Self(inner)
-    }
-
-    /// Creates expanded node ID from node ID.
-    #[must_use]
-    pub(crate) fn from_node_id(node_id: ua::NodeId) -> Self {
-        // This passes ownership into the created expanded node ID.
-        Self(unsafe { UA_EXPANDEDNODEID_NODEID(node_id.into_raw()) })
     }
 
     #[must_use]
@@ -43,6 +68,13 @@ impl ExpandedNodeId {
     #[must_use]
     pub const fn server_index(&self) -> u32 {
         self.0.serverIndex
+    }
+}
+
+impl From<ua::NodeId> for ExpandedNodeId {
+    fn from(node_id: ua::NodeId) -> Self {
+        // This passes ownership into the created expanded node ID.
+        Self(unsafe { UA_EXPANDEDNODEID_NODEID(node_id.into_raw()) })
     }
 }
 
@@ -69,7 +101,7 @@ impl str::FromStr for ExpandedNodeId {
 
         let status_code = ua::StatusCode::new({
             let str = ua::String::new(s)?;
-            // SAFETY: `UA_NodeId_parse()` expects the string passed by value but does not take
+            // SAFETY: `UA_ExpandedNodeId_parse()` expects the string passed by value but does not take
             // ownership.
             let str = unsafe { ua::String::to_raw_copy(&str) };
             unsafe { UA_ExpandedNodeId_parse(node_id.as_mut_ptr(), str) }