+++ CHANGELOG.md
@@ -19,6 +19,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add constructor `ua::Guid::new()`.
 - Add accessors for remaining `UA_NodeIdType` variants to `ua::NodeId`.
 - Add methods `null()` and `is_null()` to `ua::String` and `ua::ByteString`.
+- Add method `new()` to `ua::ExpandedNodeId`.
+- Add `From`/`Into` conversion from `ua::NodeId` to `ua::ExpandedNodeId`.
 
 ### Changed
 
@@ -26,6 +28,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Breaking: Return `Result` from `DataTypeExt::from_inner()`.
 - Breaking: Replace `ua::String::is_invalid()` with `ua::String::is_null()`.
 - Breaking: Replace `ua::ByteString::is_invalid()` with `ua::ByteString::is_null()`.
+- Breaking: Remove `into_expanded_node_id()` from `ua::NodeId`. Use `Into::into()` instead.
 
 ### Fixed